Extract PostgresqlRandomSelection and PostgresqlInOrderSelection
