+++ lib/subjects/postgresql_in_order_selection.rb
@@ -1,0 +1,15 @@
+module Subjects
+  class PostgresqlInOrderSelection
+    attr_reader :available, :limit
+
+    def initialize(available, limit)
+      @available = available
+      @limit = limit
+    end
+
+    def select
+      available.order(priority: :asc).limit(limit).pluck(:id)
+    end
+  end
+end
+
