+++ lib/subjects/postgresql_selection.rb
@@ -34,27 +34,7 @@ module Subjects
       if workflow.grouped
         query = query.where(subject_set_id: opts[:subject_set_id])
       end
-      if workflow.prioritized
-        query = query.order(priority: :asc)
-      end
       @available = query
-    end
-
-    def available_count
-      @available_count ||= available.except(:select).count
-    end
-
-    def sample(query=available)
-      direction = [:asc, :desc].sample
-      query.order(random: direction).limit(focus_set_window_size)
-    end
-
-    def focus_set_window_size
-      @focus_set_window_size ||=
-        [
-          (available_count * 0.5).ceil,
-          Panoptes::SubjectSelection.focus_set_window_size
-        ].min
     end
 
     def limit
@@ -74,24 +54,11 @@ module Subjects
     end
 
     def select_results_randomly
-      enough_available = limit < available_count
-      if enough_available
-        ids = sample.pluck(:id).sample(limit)
-        if reassign_random?
-          RandomOrderShuffleWorker.perform_async(ids)
-        end
-        ids
-      else
-        available.pluck(:id).shuffle
-      end
+      PostgresqlRandomSelection.new(available, limit).select
     end
 
     def select_results_in_order
-      available.limit(limit).pluck(:id)
-    end
-
-    def reassign_random?
-      rand < Panoptes::SubjectSelection.index_rebuild_rate
+      PostgresqlInOrderSelection.new(available, limit).select
     end
 
     def any_workflow_data_scope
